my-service/requirements.txt
numpy==1.16.2
Pillow==6.0.0
dlib==19.16.0
imutils==0.5.2
opencv-python==4.0.0.21
